Use EOS for deployment during revisions
If the OProject Git repository is used for deploying the revisions of
the document then there will be a huge number of commits that are
essentially poulluting the history of that repository. Instead, for the
bulk of the development period the CI will deploy to Feickert's web area
on EOS that is restriced to CERN users.