Restrict deployment to master (#26)
The deployment to master should only occur during a PR merge or a push
to master. This fixes an issue where a build would be triggered on
another branch that would be turned into a PR but before the PR existed
and so the control logic would still deploy to the master build location.